refactor(sra): finalize packet carrier and IO ownership (#171)
