+++ docs/internal/maintenance-runtime-slices-20260905.md
@@ -19,10 +19,40 @@ location and comments. This is a one-time move proof, not a permanent test tied 
 source layout. Existing agreement/conflict/hash/alias/repair tests remain the regression
 owner. The first focused rerun contains 131 passing tests.
 
-The core loses 198 definition lines and gains two explicit imports. This is not completion
-of every #171 checkbox: packet/input/schema generation/carrier and larger integrity
-organization remain separate follow-on slices. No same-wording conflict behavior was
-changed. Code size alone is not the acceptance criterion.
+The core loses 198 definition lines and gains two explicit imports. No same-wording
+conflict behavior was changed. Code size alone is not the acceptance criterion.
+
+## SRA slice 2 / #171 closeout
+
+The second slice moves the remaining high-cohesion runtime surfaces without changing
+judgment semantics:
+
+- `sra_packets.py` owns context admission, candidate-context weighting, state projection,
+  challenge alias construction, base/coverage/challenge/situated packets, and targeted
+  reconciliation packet construction.
+- `sra_carriers.py` owns stage prompts, output-schema dispatch surfaces, ephemeral CLI
+  command generation, observable receipt storage, context-boundary reporting, and final
+  decision copies.
+- `sra_io.py` owns the single SRA error hierarchy, JSON/JSONL persistence, run-state
+  loading, and deterministic trace-event identity.
+- `sra_runtime_core.py` is now the bounded contract-validation owner: input contract,
+  output schemas, bundle/allocation/domain invariants, and judgment validation. It no
+  longer owns packet construction, carrier orchestration, comparison normalization, or
+  runtime persistence primitives.
+
+The reconciliation packet moved with the packet owner after the first extraction exposed
+its hidden dependency on `_packet`; the focused suite failed before that correction and
+passed after the ownership boundary was fixed. Public `sra_runtime` continues to re-export
+the existing names. Runtime-integrity code imports packet and carrier owners directly,
+so it does not rely on a reverse facade import. All seven split helper modules are present
+in all five release-pack layouts.
+
+Final #171 verification on this slice: 160 focused SRA tests passed; the complete suite
+ran 1037 tests with 5 documented optional-dependency skips; Test Lifecycle remained
+77/77 valid. `sra_runtime_core.py` is 2033 lines after extraction, down from 3174 at the
+maintenance baseline. Remaining size is contract validation rather than mixed packet,
+carrier, comparison and IO ownership, so #171 stops here rather than splitting one
+validation algorithm into arbitrary small files.
 
 ## TPlan slice 1 (#172)
