+++ skills/sra/scripts/sra_runtime.py
@@ -2,5 +2,8 @@
 """Stable public facade for the version-bound SRA v0.3 runtime."""
 
 from sra_domain import *  # noqa: F401,F403
+from sra_io import *  # noqa: F401,F403
 from sra_runtime_core import *  # noqa: F401,F403
+from sra_packets import *  # noqa: F401,F403
+from sra_carriers import *  # noqa: F401,F403
 from sra_runtime_integrity import *  # noqa: F401,F403
