+++ skills/sra/scripts/sra_runtime_integrity.py
@@ -10,7 +10,9 @@ from typing import Any
 
 from sra_domain import *  # noqa: F403
 from sra_runtime_core import *  # noqa: F403
-from sra_runtime_core import _receipt_hash
+from sra_packets import *  # noqa: F403
+from sra_carriers import *  # noqa: F403
+from sra_carriers import _receipt_hash
 
 RUN_STATE_KEYS = frozenset({
     "schema_version", "run_id", "created_at", "updated_at", "mode", "view_plan",
